reduce save speed on mainfile, add prompt to overwrite backup the first time

+++ main.py
@@ -103,19 +103,18 @@ if __name__ == '__main__':
     try:
         with open('ENcycloPYdia.json', 'r') as openfile:
             enpy = json.load(openfile)
+            print("Opened main file.")
     except json.decoder.JSONDecodeError:
-        # empty JSON, lets fill:
-        # enpy = json.dumps(empty_dictionary, indent=4)
         try:
             with open('ENcycloPYdia-Backup.json', 'r') as openfile:
                 enpy = json.load(openfile)
                 print("Opened backup--main file failed to save last.")
         except Exception:
             with open("ENcycloPYdia.json", "w") as outfile:
-                # outfile.write(enpy)
                 json.dump(empty_dictionary, outfile)
-                print("Opened main file.")
+                print("Opened blank.")
 
+    overwrite_backup = False
     last_entry = ""
     last_count = 0
     try:
@@ -210,15 +209,20 @@ if __name__ == '__main__':
 
         #capture word in json
         try:
-            with open("ENcycloPYdia.json", "w") as outfile:
-                json.dump(enpy, outfile)
+            #Lets see if only saving every 5 words speeds things up a bit
+            if count % 5 == 4:
+                with open("ENcycloPYdia.json", "w") as outfile:
+                    json.dump(enpy, outfile)
         except KeyboardInterrupt:
             cprint(Fore.RED, "You interrupted the program.")
         else:
             if count % 50 == 49:
-                with open("ENcycloPYdia-Backup.json", "w") as outfile:
-                    print("SAVING BACKUP")
-                    json.dump(enpy, outfile)
+                if not overwrite_backup:
+                    if input("Do you want to overwrite the backup file? (y/n)").lower() == "y":
+                        overwrite_backup = True
+                        with open("ENcycloPYdia-Backup.json", "w") as outfile:
+                            print("SAVING BACKUP")
+                            json.dump(enpy, outfile)
 
         end_time = time.time()
         elapsed_time = end_time - start_time
